agentic_labeler/pyproject.toml
[project]
name = "angelica"
version = "0.0.1"
description = "Agentic Labeling"
authors = [
    { name = "X", email = "abcd@efgh.com" },
]
requires-python = ">=3.11,<3.13"
readme = "README.md"
#license = "License"
dependencies = [
    "joblib==1.4.2",
    "typer>=0.17.4",
    "rich==14.0.0",
    "tqdm[all]>=4.66.1,<5",
    "requests>=2.32.0,<3",
    "javalang==0.13.0",
    "toml>=0.10.2,<0.11",
    "click>=8.1.7",
    "jinja2>=3.1.2,<4",
    "openai>=1.12.0,<2",
    "python-dotenv>=1.0.1,<2",
    "tree-sitter-java>=0.23.5,<0.24",
    "seaborn>=0.13.2,<0.14",
    "matplotlib>=3.10.3,<4",
    "scikit-learn==1.8.0",
    "replicate>=1.0.7,<2",
    "cldk",
    "javabuild",
    "hamster",
    "langchain==0.3.26",
    "faiss-cpu==1.11.0",
    "langchain-community==0.3.26",
    "langgraph==0.5.4",
    "langchain-openai==0.3.30",
    "ray>=2.9.0,<2.50"
]



[project.scripts]
angelica = "angelica.cli:app"

[dependency-groups]
dev = [
    "black>=24.3.0,<25",
    "pytest>=7.4.3,<8",
    "pytest-cov>=4.1.0,<5",
    "pytest-pspec>=0.0.4,<0.0.5",
    "pylint>=3.0.2,<4",
    "flake8>=6.1.0,<7",
    "coverage>=7.3.2,<8",
    "pyinstaller>=6.6.0,<7",
    "debugpy>=1.8.2,<2",
    "ollama>=0.4.6,<0.5",
    "pip>=23,<25"
]



[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"




